+++ internal/api/handler/handler.go
@@ -316,8 +316,8 @@ func (h *DownloadsHandler) ProductMetadataHandler(c *fiber.Ctx) error {
 // @Param       p          query  string true  "Platform, valid values are returned from the `/platforms` endpoint."                                                          Example(ubuntu)
 // @Param       pv         query  string true  "Platform Version, possible values depend on the platform. For example, Ubuntu: 16.04, or 18.04 or for macOS: 10.14 or 10.15." Example(20.04)
 // @Param       m          query  string true  "Machine architecture, valid values are returned by the `/architectures` endpoint."                                            Example(x86_64)
-// @Param       v          query  string false "Version of the product to be installed. A version always takes the form `x.y.z`"                                              Default(latest)
 // @Param       pm         query  string true  "Package Manager, valid values depend on the platform (e.g., Linux: deb, tar; Windows: msi)."                                  Example(tar)
+// @Param       v          query  string false "Version of the product to be installed. A version always takes the form `x.y.z`"                                              Default(latest)
 // @Param       license_id query  string false "License ID"
 // @Param       eol        query  bool   false "EOL Products" Default(false)
 // @Success     302
@@ -363,28 +363,23 @@ func (h *DownloadsHandler) ProductDownloadHandler(c *fiber.Ctx) error {
 			buf := make([]byte, 32*1024) // 32KB buffer
 			for {
 				n, err := downloadResp.Read(buf)
+				if n > 0 {
+					if _, writeErr := w.Write(buf[:n]); writeErr != nil {
+						h.Log.Errorf("Error while streaming : %s", writeErr.Error())
+						w.Flush()
+						break
+					}
+					if err := w.Flush(); err != nil {
+						h.Log.Errorf("Error while streaming : %s", err.Error())
+						break
+					}
+				}
 				if err == io.EOF {
 					break
 				}
 				if err != nil {
 					h.Log.Errorf("Error while streaming : %s", err.Error())
-					return
-				}
-				if n > 0 {
-					if _, writeErr := w.Write(buf[:n]); writeErr != nil {
-						h.Log.Errorf("Error while streaming : %s", writeErr.Error())
-						if err := w.Flush(); err != nil {
-							h.Log.Errorf("Error while streaming : %s", err.Error())
-							break
-						}
-						return
-
-					}
-					// Explicitly flush the response
-					if err := w.Flush(); err != nil {
-						h.Log.Errorf("Error while streaming : %s", err.Error())
-						break
-					}
+					break
 				}
 			}
 			defer downloadResp.Close()
@@ -443,6 +438,7 @@ func (h *DownloadsHandler) RelatedProductsHandler(c *fiber.Ctx) error {
 // @Param       p          query    string true  "Platform, valid values are returned from the `/platforms` endpoint."                                                          Example(ubuntu)
 // @Param       pv         query    string true  "Platform Version, possible values depend on the platform. For example, Ubuntu: 16.04, or 18.04 or for macOS: 10.14 or 10.15." Example(20.04)
 // @Param       m          query    string true  "Machine architecture, valid values are returned by the `/architectures` endpoint."                                            Example(x86_64)
+// @Param       pm         query    string true  "Package Manager, valid values depend on the platform (e.g., Linux: deb, tar; Windows: msi)."								    Example(tar)
 // @Param       v          query    string false "Version of the product to be installed. A version always takes the form `x.y.z`"                                              Default(latest)
 // @Param       license_id query    string false "License ID"
 // @Success     200        {object} map[string]interface{}
@@ -610,12 +606,12 @@ func setLocals(c *fiber.Ctx) map[string]interface{} {
 		locals["valid_license"] = false
 	}
 
-	if c.Locals("request_id") != nil {
-		requestId := c.Locals("request_id").(string)
-		locals["request_id"] = requestId
+	if c.Locals("requestid") != nil {
+		requestId := c.Locals("requestid").(string)
+		locals["requestid"] = requestId
 
 	} else {
-		locals["request_id"] = ""
+		locals["requestid"] = ""
 	}
 	if c.Locals("base_url") != nil {
 		baseUrl := c.Locals("base_url").(string)